Рефактор.

+++ actions/crud/delete_book.php
@@ -11,14 +11,6 @@ elseif (!$_SESSION['is_admin'])
 
 require './configs/dbconf.php';
 
-$conn = new mysqli($servername, $username, $password, $dbname);
-
-if ($conn->connect_error) {
-    die("Connection failed: " . $conn->connect_error);
-}
-
-$conn->set_charset('utf8');
-
 $id = $conn->real_escape_string($_GET['id']);
 
 $sql = 'SELECT cover FROM ' . $dbname . ' .books WHERE book_id = ' . $id .';';
@@ -27,7 +19,10 @@ $result = $conn->query($sql);
 if ($result->num_rows > 0)
 {
     $row = $result->fetch_assoc();
-    unlink($row['cover']);   
+    if(file_exists($row['cover']))
+    {
+        unlink($row['cover']);   
+    }
 }
 
 $sql = 'DELETE FROM ' . $dbname . '.books WHERE book_id = ' . $id;
